pyproject.toml
[tool.poetry]
name = "essential_sqlalchemy"
version = "0.0.1"
description = "A short description of the project"
authors = [
    "Vamsi Ampolu <vamsiampolu3000@gmail.com>",
]
license = "MIT"
readme = "README.md"

documentation = "https://vamsiampolu.github.io/essential_sqlalchemy"
homepage = "https://vamsiampolu.github.io/essential_sqlalchemy"
repository = "https://github.com/vamsiampolu/essential_sqlalchemy"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]

[tool.poetry.dependencies]
python = ">=3.8, <4.0"
sqlalchemy = {extras = ["mypy"], version = "^1.4.45"}
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
pydantic-sqlalchemy = "^0.0.9"

[tool.poetry.dev-dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-eradicate = "*"
flake8-logging-format = "*"
isort = "*"
mkdocstrings = {version = ">=0.18", extras = ["python"]}
mkdocs-material = "*"
mypy = "*"
pep8-naming = "*"
pre-commit = "*"
pymdown-extensions = "*"
pytest = "*"
pytest-github-actions-annotate-failures = "*"
pytest-cov = "*"
python-kacl = "*"
pyupgrade = "*"
tryceratops = "*"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.16.5"
mypy = "^0.991"
pre-commit = "^3.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.black]
target-version = [ "py38", "py39"]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = """\
    --cov essential_sqlalchemy \
    --cov tests \
    --cov-report term-missing \
    --no-cov-on-fail \
"""


[tool.coverage.report]
fail_under = 100
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover'
]

[tool.mypy]
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
show_error_codes = true


[tool.poe.tasks]
dev = "poetry run python -m essential_sqlalchemy"
build = "poetry build"
format = "black ./"
type_lint="mypy essential_sqlalchemy"
